Fixed caching, functionality to index PDF-files
Caching does not work, at least not in Sitecore 7.5. The only text that was printed from cache is "System.Xml.XmlTextWriter". Caching is fixed using a custom StringWriter and also added code to index files, PDF is set as example. Added code to only index files that are viewable by anonymous users